ToStringでノーツ名が返る件を修正

+++ TJA-Supporter.Lib/Note.cs
@@ -82,7 +82,7 @@ namespace TJA_Supporter.Lib
         #region Method
         public override string ToString()
         {
-            return Type.ToString();
+            return ((int)Type).ToString();
         }
         #endregion
